refactor: move gene to different folders
